sample_services/kafka-python-flow/requirement.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
confluent-kafka==2.5.3
exceptiongroup==1.2.2
fastapi==0.112.2
h11==0.14.0
idna==3.8
kafka-python==2.0.2
numpy==2.1.1
pandas==2.2.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.4
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
